pyproject.toml
[tool.poetry]
name = "fun_relations"
version = "0.1.0"
description = "FunRelations mobile/web api"
authors = ["TomohiroHiratsuka <info@easpe.jp>"]

[tool.poetry.dependencies]
python = "^3.11.3" # Updated
fastapi = "^0.95.2" # Updated
uvicorn = "^0.22.0" # Updated
SQLAlchemy = "^2.0.15" # Updated
python-dotenv = "^0.21.0" # Updated
alembic = "^1.11.1"
sqlalchemy-utils = "^0.41.1"
aiomysql = "^0.1.1"
cryptography = "^41.0.0" # Updated
flake8 = "^6.0.0"
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120

[tool.mypy]
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_optional = true
